Change api responses

+++ src/main/java/com/legal/food/api/model/PantryItem.java
@@ -23,8 +23,8 @@ public class PantryItem {
     @Column(name = "pantry_name")
     private String name;
     private String barCode;
-    private Integer placeDate;
-    private Integer expireDate;
+    private Long placeDate;
+    private Long expireDate;
     private Float quantity;
     private String unit;
     @ManyToOne(targetEntity = Ingredient.class)
